updated phaser version removed unused boilerplate limited fps to 48
