--- config/karma.conf.js
@@ -1,37 +1,0 @@
-var webpackConfig = require('./webpack.test');
-
-module.exports = function (config) {
-    var _config = {
-        basePath: '',
-
-        frameworks: ['jasmine'],
-
-        files: [
-            {pattern: './config/karma-test-shim.js', watched: false}
-        ],
-
-        preprocessors: {
-            './config/karma-test-shim.js': ['webpack', 'sourcemap']
-        },
-
-        webpack: webpackConfig,
-
-        webpackMiddleware: {
-            stats: 'errors-only'
-        },
-
-        webpackServer: {
-            noInfo: true
-        },
-
-        reporters: ['progress'],
-        port: 9876,
-        colors: true,
-        logLevel: config.LOG_INFO,
-        autoWatch: false,
-        browsers: ['PhantomJS'],
-        singleRun: true
-    };
-
-    config.set(_config);
-};
