--- config/webpack.common.js
@@ -1,80 +1,0 @@
-var webpack = require('webpack');
-var HtmlWebpackPlugin = require('html-webpack-plugin');
-var ExtractTextPlugin = require('extract-text-webpack-plugin');
-var helpers = require('./helpers');
-var path = require('path');
-
-var phaserModule = path.join(__dirname, '../node_modules/phaser/');
-var phaser = path.join(phaserModule, 'build/custom/phaser-split.js');
-var pixi = path.join(phaserModule, 'build/pixi.js');
-var p2 = path.join(phaserModule, 'build/p2.js');
-
-module.exports = {
-    entry: {
-        'polyfills': './src/polyfills.ts',
-        'vendor': './src/vendor.ts',
-        'app': './src/main.ts'
-    },
-
-    resolve: {
-        extensions: ['.js', '.ts'],
-        alias: {
-            'phaser': phaser,
-            'pixi': pixi,
-            'p2': p2
-        }
-    },
-
-    module: {
-        rules: [
-            {
-                test: /pixi\.js/,
-                loader: 'expose-loader?PIXI'
-            },
-            {
-                test: /phaser-split\.js$/,
-                loader: 'expose-loader?Phaser'
-            },
-            {
-                test: /p2\.js/,
-                loader: 'expose-loader?p2'
-            },
-            {
-                test: /\.ts$/,
-                exclude: helpers.root('node_modules'),
-                use: ['awesome-typescript-loader', 'angular2-template-loader']
-            },
-            {
-                test: /\.html$/,
-                loader: 'html-loader'
-            },
-            {
-                test: /\.(png|jpe?g|gif|svg|woff|woff2|ttf|eot|ico)$/,
-                loader: 'file-loader?name=assets/[name].[hash].[ext]'
-            },
-            {
-                test: /\.css$/,
-                exclude: helpers.root('src', 'app'),
-                use: ExtractTextPlugin.extract({
-                    fallback: "style-loader",
-                    use: "css-loader"
-                })
-            },
-            {
-                test: /\.css$/,
-                include: helpers.root('src', 'app'),
-                loader: 'raw-loader'
-            }
-        ]
-    },
-
-    plugins: [
-        new webpack.optimize.CommonsChunkPlugin({
-            name: ['app', 'vendor', 'polyfills']
-        }),
-
-        new HtmlWebpackPlugin({
-            template: 'src/index.html'
-        })
-    ]
-};
