--- config/webpack.dev.js
@@ -1,24 +1,0 @@
-var webpackMerge = require('webpack-merge');
-var ExtractTextPlugin = require('extract-text-webpack-plugin');
-var commonConfig = require('./webpack.common.js');
-var helpers = require('./helpers');
-
-module.exports = webpackMerge(commonConfig, {
-    devtool: 'cheap-module-eval-source-map',
-
-    output: {
-        path: helpers.root('dist'),
-        publicPath: 'http://localhost:3000/',
-        filename: '[name].js',
-        chunkFilename: '[id].chunk.js'
-    },
-
-    plugins: [
-        new ExtractTextPlugin('[name].css')
-    ],
-
-    devServer: {
-        historyApiFallback: true,
-        stats: 'minimal'
-    }
-});
