--- config/webpack.test.js
@@ -1,37 +1,0 @@
-var helpers = require('./helpers');
-
-module.exports = {
-    devtool: 'inline-source-map',
-
-    resolve: {
-        extensions: ['.ts', '.js']
-    },
-
-    module: {
-        rules: [
-            {
-                test: /\.ts$/,
-                use: ['awesome-typescript-loader', 'angular2-template-loader']
-            },
-            {
-                test: /\.html$/,
-                loader: 'html-loader'
-
-            },
-            {
-                test: /\.(png|jpe?g|gif|svg|woff|woff2|ttf|eot|ico)$/,
-                loader: 'null-loader'
-            },
-            {
-                test: /\.css$/,
-                exclude: helpers.root('src', 'app'),
-                loader: 'null-loader'
-            },
-            {
-                test: /\.css$/,
-                include: helpers.root('src', 'app'),
-                loader: 'raw-loader'
-            }
-        ]
-    }
-}
