--- karma.conf.js
@@ -1,1 +1,0 @@
-module.exports = require('./config/karma.conf.js');
