+++ package.json
@@ -14,31 +14,18 @@
     "url": "git+https://github.com/wirus15/bombajs.git"
   },
   "dependencies": {
-    "@angular/common": "^2.1.0",
-    "@angular/compiler": "^2.1.0",
-    "@angular/core": "^2.1.0",
-    "@angular/forms": "^2.1.0",
-    "@angular/http": "^2.1.0",
-    "@angular/platform-browser": "^2.1.0",
-    "@angular/platform-browser-dynamic": "^2.1.0",
-    "@angular/router": "^3.1.0",
-    "@angular/upgrade": "^2.1.0",
     "core-js": "^2.4.1",
     "es6-promisify": "^5.0.0",
     "es6-shim": "^0.35.3",
     "huject": "^1.3.4",
     "lodash": "^4.17.4",
-    "phaser": "https://github.com/jamesgroat/phaser/tarball/typings",
-    "reflect-metadata": "^0.1.9",
-    "rxjs": "^5.1.1",
-    "systemjs": "^0.20.9",
-    "zone.js": "^0.7.7"
+    "phaser-ce": "^2.8.1",
+    "reflect-metadata": "^0.1.9"
   },
   "devDependencies": {
     "@types/core-js": "^0.9.35",
     "@types/jasmine": "^2.5.35",
     "@types/node": "^6.0.45",
-    "angular2-template-loader": "^0.5.0",
     "awesome-typescript-loader": "^3.0.4",
     "css-loader": "^0.25.0",
     "expose-loader": "^0.7.1",
