--- src/app/app.component.html
@@ -1,4 +1,0 @@
-<main>
-    <h1>Hello from Angular App with Webpack</h1>
-    <img src="../../public/images/angular.png">
-</main>
