--- src/app/app.component.ts
@@ -1,9 +1,0 @@
-import { Component } from '@angular/core';
-import '../../public/css/styles.css';
-
-@Component({
-    selector: 'my-app',
-    template: '<router-outlet></router-outlet>',
-    styleUrls: ['./app.component.css']
-})
-export class AppComponent { }
