+++ src/app/game/game_events.ts
@@ -1,4 +1,3 @@
-import * as Phaser from "phaser";
 import {ConstructorInject} from 'huject';
 
 @ConstructorInject
