--- src/app/menu/menu.component.html
@@ -1,11 +1,0 @@
-<ul>
-    <li>
-        <a routerLink="/game">Start</a>
-    </li>
-    <li>
-        <a href="#">Opcje</a>
-    </li>
-    <li>
-        <a href="#">Instrukcje</a>
-    </li>
-</ul>
