--- src/app/menu/menu.component.ts
@@ -1,7 +1,0 @@
-import {Component} from "@angular/core";
-
-@Component({
-    selector: 'b-menu',
-    templateUrl: './menu.component.html'
-})
-export class MenuComponent { }
