+++ src/index.html
@@ -8,6 +8,5 @@
 </head>
 <body>
 <div id="gameCanvas"></div>
-<script src="main.ts"></script>
 </body>
 </html>
