+++ src/polyfills.ts
@@ -2,9 +2,6 @@ import 'core-js/es6';
 import 'core-js/es7/reflect';
 import 'reflect-metadata';
 
-require('zone.js/dist/zone');
-
 if (process.env.ENV !== 'production') {
     Error['stackTraceLimit'] = Infinity;
-    require('zone.js/dist/long-stack-trace-zone');
 }
