+++ src/vendor.ts
@@ -1,12 +1,5 @@
-/// <reference path="../node_modules/phaser/typescript/phaser.comments.d.ts"/>
+/// <reference path="../node_modules/phaser-ce/typescript/phaser.comments.d.ts"/>
 
-import '@angular/platform-browser';
-import '@angular/platform-browser-dynamic';
-import '@angular/core';
-import '@angular/common';
-import '@angular/http';
-import '@angular/router';
-import 'rxjs';
 import 'pixi';
 import 'p2';
-import 'phaser';
+import 'phaser-ce';
